# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Dokumentasi API Barantin",
        "description": "Data Request",
        "contact": {
            "email": "nandaputraservice@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1",
            "description": "Api Admin Barantin"
        },
        {
            "url": "http://localhost:8000/api/v2",
            "description": "Api User Barantin Test"
        }
    ],
    "paths": {
...
        "/barantin/perusahaan/induk/{take}": {
            "get": {
                "tags": [
                    "Barantin Admin"
                ],
                "summary": "Dapatkan Data Barantin Perusahaan Induk",
                "description": "Mengambil data Barantin Perusahaan Induk menggunakan parameter take",
                "operationId": "getAllDataBarantinPerusahaanInduk",
                "parameters": [
                    {
                        "name": "take",
                        "in": "path",
                        "description": "Jumlah data yang akan diambil",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "barantin data by id"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "nama": {
                                                    "type": "string",
                                                    "example": "John Doe"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "john.doe@example.com"
                                                },
                                                "no_hp": {
                                                    "type": "string",
                                                    "example": "08123456789"
                                                },
                                                "alamat": {
                                                    "type": "string",
                                                    "example": "Jl. Sudirman No. 123"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "example": "2023-02-28 12:34:56"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "example": "2023-02-28 12:34:56"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Data Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "data not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/barantin/perusahaan/cabang/{take}": {
            "get": {
                "tags": [
                    "Barantin Admin"
                ],
                "summary": "Dapatkan Data Barantin Perusahaan Cabang",
                "description": "Mengambil data Barantin Perusahaan Cabang",
                "operationId": "getAllDataBarantinPerusahaanCabang",
                "parameters": [
                    {
                        "name": "take",
                        "in": "path",
                        "description": "Jumlah data yang ingin diambil",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "barantin data perusahaan cabang"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "nama": {
                                                        "type": "string",
                                                        "example": "John Doe"
                                                    },
                                                    "email": {
                                                        "type": "string",
                                                        "example": "john.doe@example.com"
                                                    },
                                                    "no_hp": {
                                                        "type": "string",
                                                        "example": "08123456789"
                                                    },
                                                    "alamat": {
                                                        "type": "string",
                                                        "example": "Jl. Sudirman No. 123"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "example": "2023-02-28 12:34:56"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "example": "2023-02-28 12:34:56"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Data Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "data not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
        "/barantin/{register_id}/register": {
            "get": {
                "tags": [
                    "Barantin Admin"
                ],
                "summary": "Get Barantin Data By Register ID",
                "description": "Get Barantin Data By Register ID",
                "operationId": "d60500c4bf47725ade56f02e7675ce5f",
                "parameters": [
                    {
                        "name": "register_id",
                        "in": "path",
                        "description": "Register ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Data Not Found"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearer_token": {
                "type": "apiKey",
                "description": "Tambahkan Kata \"Bearer\" Sebelum token contoh: (Bearer xx|497zSWISvECINqxxxxxxxxxxxxxxx)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}